refactor names and move to www folder all www pages
