--- style.css
@@ -1,4 +1,0 @@
-body{
-    color: honeydew;
-    background-color: darkslategray;
-}
