+++ www/index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
+    <link rel="shortcut icon" href="favicon.ico">
     <link rel="stylesheet" href="style.css">
 </head>
 <body>
@@ -11,9 +12,10 @@
     <div>Canvas</div>
     <canvas id="canvas">
     </canvas>
+    <a href="./form.html">forma</a>
 
     <div>Obrazek</div>
-    <img src="small.png" alt="terraria home">
+    <img src="big.png" alt="terraria home">
     
     <script src="sc.js"></script>
 </body>
