+++ www/style.css
@@ -1,0 +1,30 @@
+body{
+    color: honeydew;
+    background-color: darkslategray;
+}
+
+.gold input[type=text]:focus {
+    background-color: gold;
+}
+
+input[type=text]:focus {
+    background-color: lightblue;
+}
+
+input[type=text] {
+    width: 100%;
+    padding: 12px 20px;
+    margin: 8px 0;
+    box-sizing: border-box;
+  }
+
+input[type=button], input[type=submit], input[type=reset] {
+    background-color: #4CAF50;
+    border: none;
+    color: white;
+    padding: 16px 32px;
+    text-decoration: none;
+    margin: 4px 2px;
+    cursor: pointer;
+}
+
